unify itoa and dtoa in one file "internal/number"
