+++ std/assembly/internal/number.ts
@@ -1,9 +1,3 @@
-/*
- * Implementation based on dtoa-pldi2010).pdf
- * "Printing Floating-Point Numbers Quickly and Accurately with Integers" by Florian Loitsh paper and implementation
- * Also some parts inspired from milo's implementation:
- * https://github.com/miloyip/dtoa-benchmark/blob/master/src/milo/dtoa_milo.h
- */
 
 import {
   CharCode,
@@ -16,11 +10,63 @@ import {
   loadUnsafe
 } from "./arraybuffer";
 
-import {
-  POWERS10,
-  decimalCount32,
-  utoa32_core
-} from "./itoa";
+@inline
+export function POWERS10(): u32[] {
+  const table: u32[] = [
+    1,
+    10,
+    100,
+    1000,
+    10000,
+    100000,
+    1000000,
+    10000000,
+    100000000,
+    1000000000
+  ];
+  return table; // inlines to a constant memory offset
+}
+
+/*
+  Lookup table for pairwise char codes in range [0-99]
+
+  "00", "01", "02", "03", "04", "05", "06", "07", "08", "09",
+  "10", "11", "12", "13", "14", "15", "16", "17", "18", "19",
+  "20", "21", "22", "23", "24", "25", "26", "27", "28", "29",
+  "30", "31", "32", "33", "34", "35", "36", "37", "38", "39",
+  "40", "41", "42", "43", "44", "45", "46", "47", "48", "49",
+  "50", "51", "52", "53", "54", "55", "56", "57", "58", "59",
+  "60", "61", "62", "63", "64", "65", "66", "67", "68", "69",
+  "70", "71", "72", "73", "74", "75", "76", "77", "78", "79",
+  "80", "81", "82", "83", "84", "85", "86", "87", "88", "89",
+  "90", "91", "92", "93", "94", "95", "96", "97", "98", "99"
+*/
+@inline
+export function DIGITS(): u32[] {
+  const table: u32[] = [
+    0x00300030, 0x00310030, 0x00320030, 0x00330030, 0x00340030,
+    0x00350030, 0x00360030, 0x00370030, 0x00380030, 0x00390030,
+    0x00300031, 0x00310031, 0x00320031, 0x00330031, 0x00340031,
+    0x00350031, 0x00360031, 0x00370031, 0x00380031, 0x00390031,
+    0x00300032, 0x00310032, 0x00320032, 0x00330032, 0x00340032,
+    0x00350032, 0x00360032, 0x00370032, 0x00380032, 0x00390032,
+    0x00300033, 0x00310033, 0x00320033, 0x00330033, 0x00340033,
+    0x00350033, 0x00360033, 0x00370033, 0x00380033, 0x00390033,
+    0x00300034, 0x00310034, 0x00320034, 0x00330034, 0x00340034,
+    0x00350034, 0x00360034, 0x00370034, 0x00380034, 0x00390034,
+    0x00300035, 0x00310035, 0x00320035, 0x00330035, 0x00340035,
+    0x00350035, 0x00360035, 0x00370035, 0x00380035, 0x00390035,
+    0x00300036, 0x00310036, 0x00320036, 0x00330036, 0x00340036,
+    0x00350036, 0x00360036, 0x00370036, 0x00380036, 0x00390036,
+    0x00300037, 0x00310037, 0x00320037, 0x00330037, 0x00340037,
+    0x00350037, 0x00360037, 0x00370037, 0x00380037, 0x00390037,
+    0x00300038, 0x00310038, 0x00320038, 0x00330038, 0x00340038,
+    0x00350038, 0x00360038, 0x00370038, 0x00380038, 0x00390038,
+    0x00300039, 0x00310039, 0x00320039, 0x00330039, 0x00340039,
+    0x00350039, 0x00360039, 0x00370039, 0x00380039, 0x00390039
+  ];
+  return table; // inlines to a constant memory offset
+}
 
 @inline
 function EXP_POWERS(): i16[] {
@@ -66,6 +112,249 @@ function FRC_POWERS(): u64[] {
     0x9E19DB92B4E31BA9, 0xEB96BF6EBADF77D9, 0xAF87023B9BF0EE6B
   ];
   return table;
+}
+
+// Count number of decimals for u32 values
+// In our case input value always non-zero so we can simplify some parts
+export function decimalCount32(value: u32): u32 {
+  if (ASC_SHRINK_LEVEL >= 1) {
+    let l: u32 = 32 - clz<u32>(value); // log2
+    let t = l * 1233 >>> 12; // log10
+
+    let lutbuf = <ArrayBuffer>POWERS10().buffer_;
+    let power  = loadUnsafe<u32,u32>(lutbuf, t);
+    t -= <u32>(value < power);
+    return t + 1;
+  } else {
+    if (value < 100000) {
+      if (value < 100) {
+        return select<u32>(1, 2, value < 10);
+      } else {
+        let m = select<u32>(4, 5, value < 10000);
+        return select<u32>(3, m, value < 1000);
+      }
+    } else {
+      if (value < 10000000) {
+        return select<u32>(6, 7, value < 1000000);
+      } else {
+        let m = select<u32>(9, 10, value < 1000000000);
+        return select<u32>(8, m, value < 100000000);
+      }
+    }
+  }
+}
+
+// Count number of decimals for u64 values
+// In our case input value always greater than 2^32-1 so we can skip some parts
+export function decimalCount64(value: u64): u32 {
+  if (ASC_SHRINK_LEVEL >= 1) {
+    let l: u32 = 64 - <u32>clz<u64>(value); // log2
+    let t = l * 1233 >>> 12; // log10
+
+    let lutbuf = <ArrayBuffer>POWERS10().buffer_;
+    let power  = loadUnsafe<u32,u64>(lutbuf, t - 10);
+    t -= <u32>(value < 10000000000 * power);
+    return t + 1;
+  } else {
+    if (value < 1000000000000000) {
+      if (value < 1000000000000) {
+        return select<u32>(11, 12, value < 100000000000);
+      } else {
+        let m = select<u32>(14, 15, value < 100000000000000);
+        return select<u32>(13, m, value < 10000000000000);
+      }
+    } else {
+      if (value < 100000000000000000) {
+        return select<u32>(16, 17, value < 10000000000000000);
+      } else {
+        let m = select<u32>(19, 20, value < 10000000000000000000);
+        return select<u32>(18, m, value < 1000000000000000000);
+      }
+    }
+  }
+}
+
+function utoa32_lut(buffer: usize, num: u32, offset: usize): void {
+  var lutbuf = <ArrayBuffer>DIGITS().buffer_;
+
+  while (num >= 10000) {
+    // in most VMs i32/u32 div and modulo by constant can be shared and simplificate
+    let t = num / 10000;
+    let r = num % 10000;
+    num = t;
+
+    let d1 = r / 100;
+    let d2 = r % 100;
+
+    let digits1 = loadUnsafe<u32,u64>(lutbuf, d1);
+    let digits2 = loadUnsafe<u32,u64>(lutbuf, d2);
+
+    offset -= 4;
+    store<u64>(buffer + (offset << 1), digits1 | (digits2 << 32), STRING_HEADER_SIZE);
+  }
+
+  if (num >= 100) {
+    let t  = num / 100;
+    let d1 = num % 100;
+    num = t;
+    offset -= 2;
+    let digits = loadUnsafe<u32,u32>(lutbuf, d1);
+    store<u32>(buffer + (offset << 1), digits, STRING_HEADER_SIZE);
+  }
+
+  if (num >= 10) {
+    offset -= 2;
+    let digits = loadUnsafe<u32,u32>(lutbuf, num);
+    store<u32>(buffer + (offset << 1), digits, STRING_HEADER_SIZE);
+  } else {
+    offset -= 1;
+    let digit = CharCode._0 + num;
+    store<u16>(buffer + (offset << 1), digit, STRING_HEADER_SIZE);
+  }
+}
+
+function utoa64_lut(buffer: usize, num: u64, offset: usize): void {
+  var lutbuf = <ArrayBuffer>DIGITS().buffer_;
+
+  while (num >= 100000000) {
+    let t = num / 100000000;
+    let r = <usize>(num - t * 100000000);
+    num = t;
+
+    let b = r / 10000;
+    let c = r % 10000;
+
+    let b1 = b / 100;
+    let b2 = b % 100;
+    let c1 = c / 100;
+    let c2 = c % 100;
+
+    let digits1 = loadUnsafe<u32,u64>(lutbuf, c1);
+    let digits2 = loadUnsafe<u32,u64>(lutbuf, c2);
+
+    offset -= 4;
+    store<u64>(buffer + (offset << 1), digits1 | (digits2 << 32), STRING_HEADER_SIZE);
+
+    digits1 = loadUnsafe<u32,u64>(lutbuf, b1);
+    digits2 = loadUnsafe<u32,u64>(lutbuf, b2);
+
+    offset -= 4;
+    store<u64>(buffer + (offset << 1), digits1 | (digits2 << 32), STRING_HEADER_SIZE);
+  }
+
+  utoa32_lut(buffer, <u32>num, offset);
+}
+
+function utoa_simple<T>(buffer: usize, num: T, offset: usize): void {
+  do {
+    let t = num / 10;
+    let r = <u32>(num % 10);
+    num = t;
+    offset -= 1;
+    store<u16>(buffer + (offset << 1), CharCode._0 + r, STRING_HEADER_SIZE);
+  } while (num);
+}
+
+@inline
+export function utoa32_core(buffer: usize, num: u32, offset: u32): void {
+  if (ASC_SHRINK_LEVEL >= 1) {
+    utoa_simple(buffer, num, offset);
+  } else {
+    utoa32_lut(buffer, num, offset);
+  }
+}
+
+@inline
+export function utoa64_core(buffer: usize, num: u64, offset: u32): void {
+  if (ASC_SHRINK_LEVEL >= 1) {
+    utoa_simple(buffer, num, offset);
+  } else {
+    utoa64_lut(buffer, num, offset);
+  }
+}
+
+export function utoa32(value: u32): String {
+  if (!value) return "0";
+
+  var decimals = decimalCount32(value);
+  var buffer   = allocateUnsafeString(decimals);
+
+  utoa32_core(changetype<usize>(buffer), value, decimals);
+  return buffer;
+}
+
+export function itoa32(value: i32): String {
+  if (!value) return "0";
+
+  var sign = value < 0;
+  if (sign) value = -value;
+
+  var decimals = decimalCount32(value) + <u32>sign;
+  var buffer   = allocateUnsafeString(decimals);
+
+  utoa32_core(changetype<usize>(buffer), value, decimals);
+  if (sign) store<u16>(changetype<usize>(buffer), CharCode.MINUS, STRING_HEADER_SIZE);
+
+  return buffer;
+}
+
+export function utoa64(value: u64): String {
+  if (!value) return "0";
+
+  var buffer: String;
+  if (value <= u32.MAX_VALUE) {
+    let val32    = <u32>value;
+    let decimals = decimalCount32(val32);
+    buffer = allocateUnsafeString(decimals);
+    utoa32_core(changetype<usize>(buffer), val32, decimals);
+  } else {
+    let decimals = decimalCount64(value);
+    buffer = allocateUnsafeString(decimals);
+    utoa64_core(changetype<usize>(buffer), value, decimals);
+  }
+  return buffer;
+}
+
+export function itoa64(value: i64): String {
+  if (!value) return "0";
+
+  var sign = value < 0;
+  if (sign) value = -value;
+
+  var buffer: String;
+  if (<u64>value <= <u64>u32.MAX_VALUE) {
+    let val32    = <u32>value;
+    let decimals = decimalCount32(val32) + <u32>sign;
+    buffer = allocateUnsafeString(decimals);
+    utoa32_core(changetype<usize>(buffer), val32, decimals);
+  } else {
+    let decimals = decimalCount64(value) + <u32>sign;
+    buffer = allocateUnsafeString(decimals);
+    utoa64_core(changetype<usize>(buffer), value, decimals);
+  }
+  if (sign) store<u16>(changetype<usize>(buffer), CharCode.MINUS, STRING_HEADER_SIZE);
+
+  return buffer;
+}
+
+export function itoa<T>(value: T): String {
+  if (!isInteger<T>()) {
+    assert(false); // unexpecteble non-integer generic type
+  } else {
+    if (isSigned<T>()) {
+      if (sizeof<T>() <= 4) {
+        return itoa32(<i32>value);
+      } else {
+        return itoa64(<i64>value);
+      }
+    } else {
+      if (sizeof<T>() <= 4) {
+        return utoa32(<u32>value);
+      } else {
+        return utoa64(<u64>value);
+      }
+    }
+  }
 }
 
 var _K: i32 = 0;
