+++ tests/compiler/std/array-literal.untouched.wat
@@ -637,7 +637,7 @@
     (call $~lib/env/abort
      (i32.const 0)
      (i32.const 136)
-     (i32.const 26)
+     (i32.const 25)
      (i32.const 39)
     )
     (unreachable)
@@ -737,7 +737,7 @@
     (call $~lib/env/abort
      (i32.const 0)
      (i32.const 136)
-     (i32.const 26)
+     (i32.const 25)
      (i32.const 39)
     )
     (unreachable)
