+++ tests/compiler/std/gc-array.optimized.wat
@@ -2774,7 +2774,7 @@
       (call $~lib/env/abort
        (i32.const 0)
        (i32.const 72)
-       (i32.const 90)
+       (i32.const 89)
        (i32.const 41)
       )
       (unreachable)
