+++ tests/compiler/std/gc-array.untouched.wat
@@ -3478,7 +3478,7 @@
       (call $~lib/env/abort
        (i32.const 0)
        (i32.const 72)
-       (i32.const 90)
+       (i32.const 89)
        (i32.const 41)
       )
       (unreachable)
